requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
channels==2.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-renderer-xlsx==0.4.0
et-xmlfile==1.1.0
future==0.18.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
mypy-extensions==0.4.3
netmiko==3.4.0
nornir==3.1.1
nornir-netmiko==0.1.1
nornir-utils==0.1.2
ntc-templates==2.0.0
openpyxl==3.0.7
paramiko==2.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyserial==3.5
python-gitlab==2.8.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scp==0.13.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tenacity==7.0.0
textfsm==1.1.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
zope.interface==5.4.0
